requirements.txt
discord.py[voice]>=2.3.1
python-dotenv
openai
PyNaCl
